modification des fichier avant fusion
